ZERO warinings :)

+++ Eternal/src/Eternal/ImGui/ImGuiLayer.cpp
@@ -1,10 +1,10 @@
 #include "ETpch.h"
 
-#include "ImGuiLayer.h"
+#include "Eternal/ImGui/ImGuiLayer.h"
 
-#include "imgui.h"
-#include "examples/imgui_impl_glfw.cpp"
-#include "examples/imgui_impl_opengl3.cpp"
+#include <imgui.h>
+#include <examples/imgui_impl_glfw.h>
+#include <examples/imgui_impl_opengl3.h>
 
 #include "Eternal/Application.h"
 
@@ -81,7 +81,7 @@ namespace Eternal {
 		ImGuiIO& io = ImGui::GetIO();
 
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 		
 		// Rendering
 		ImGui::Render();
