+++ Eternal/src/Platform/Windows/WindowsInput.h
@@ -10,8 +10,8 @@ namespace Eternal {
 		virtual bool IsKeyPressedImpl(int keycode) override;
 
 		virtual bool IsMouseButtonPressedImpl(int button) override;
-		virtual std::pair<float, float> GetMousePositionImpl() override;
-		virtual float GetMouseXImpl() override;
-		virtual float GetMouseYImpl() override;
+		virtual std::pair<double, double> GetMousePositionImpl() override;
+		virtual double GetMouseXImpl() override;
+		virtual double GetMouseYImpl() override;
 	};
 }
